requirements.txt
python-telegram-bot>=20.0
llama-cpp-python
python-dotenv
